Removed Ramped and renamed
