+++ geneticengine/core/representations/grammatical_evolution/ge.py
@@ -8,10 +8,8 @@ from typing import Tuple
 from typing import Type
 
 from geneticengine.core.grammar import Grammar
-from geneticengine.core.random.sources import RandomSource
 from geneticengine.core.random.sources import Source
 from geneticengine.core.representations.api import Representation
-from geneticengine.core.representations.tree.treebased import Grow
 from geneticengine.core.representations.tree.treebased import random_node
 from geneticengine.core.tree import TreeNode
 
@@ -67,7 +65,7 @@ class ListWrapper(Source):
 
 def create_tree(g: Grammar, ind: Genotype, depth: int) -> TreeNode:
     rand: Source = ListWrapper(ind.dna)
-    return random_node(rand, g, depth, g.starting_symbol, method=Grow)
+    return random_node(rand, g, depth, g.starting_symbol)
 
 
 class GrammaticalEvolutionRepresentation(Representation[Genotype]):
