+++ geneticengine/core/representations/grammatical_evolution/structured_ge.py
@@ -12,7 +12,6 @@ from geneticengine.core.grammar import Grammar
 from geneticengine.core.random.sources import RandomSource
 from geneticengine.core.random.sources import Source
 from geneticengine.core.representations.api import Representation
-from geneticengine.core.representations.tree.treebased import Grow
 from geneticengine.core.representations.tree.treebased import random_node
 from geneticengine.core.tree import TreeNode
 from geneticengine.core.utils import get_arguments
@@ -105,7 +104,7 @@ class StructuredListWrapper(Source):
 
 def create_tree(g: Grammar, ind: Genotype, depth: int) -> TreeNode:
     rand: Source = StructuredListWrapper(ind.dna)
-    return random_node(rand, g, depth, g.starting_symbol, method=Grow)
+    return random_node(rand, g, depth, g.starting_symbol)
 
 
 class StructuredGrammaticalEvolutionRepresentation(Representation[Genotype]):
