+++ geneticengine/core/representations/tree/smt.py
@@ -1,0 +1,85 @@
+from __future__ import annotations
+
+from typing import Any
+from typing import Callable
+from typing import Dict
+from typing import List
+from typing import Optional
+from typing import Tuple
+from typing import Type
+from typing import TypeVar
+
+import z3
+
+# TODO: make non static
+class SMTResolver:
+    clauses: list[Any] = []
+    receivers: dict[str, Callable] = {}
+    types: dict[str, Callable] = {}
+
+    @staticmethod
+    def add_clause(claus, recs: dict[str, Callable]):
+        SMTResolver.clauses.extend(claus)
+        for k, v in recs.items():
+            SMTResolver.receivers[k] = v
+
+    @staticmethod
+    def register_type(name, typ):
+        SMTResolver.types[name] = SMTResolver.to_z3_typ(typ)
+
+    @staticmethod
+    def to_z3_typ(typ):
+        return z3.Bool if typ == bool else z3.Int if typ == int else z3.Real
+
+    @staticmethod
+    def resolve_clauses():
+
+        if not SMTResolver.receivers:
+            return  # don't try to smt solve if we don't need to
+
+        solver = z3.Solver()
+
+        solver.set(":random-seed", 1)
+        solver.reset()
+
+        for clause in SMTResolver.clauses:
+            solver.add(clause(SMTResolver.types))
+        res = solver.check()
+
+        if res != z3.sat:
+            raise Exception(f"{solver} failed with {res}")
+
+        model = solver.model()
+        for (name, recv) in SMTResolver.receivers.items():
+            evaled = model.eval(
+                SMTResolver.types[name](
+                    name,
+                ),
+                model_completion=True,
+            )
+
+            recv(SMTResolver.get_type(evaled))
+
+        SMTResolver.clauses = []
+        SMTResolver.receivers = {}
+        SMTResolver.types = {}
+
+    @staticmethod
+    def get_type(evaled):
+        if type(evaled) == z3.z3.BoolRef:
+            evaled = bool(str(evaled))
+        elif type(evaled) == z3.z3.IntNumRef:
+            evaled = int(str(evaled))
+        elif type(evaled) == z3.z3.RatNumRef:
+            evaled = eval(str(evaled))
+        else:
+            raise NotImplementedError(
+                f"Don't know what to do with {type(evaled)} {evaled}",
+            )
+        return evaled
+
+    @staticmethod
+    def register_const(ident, val):
+        SMTResolver.register_type(ident, type(val))
+        ty = SMTResolver.types[ident]
+        SMTResolver.clauses.append(lambda _: ty(ident) == val)
