+++ geneticengine/core/representations/tree/treebased.py
@@ -3,317 +3,25 @@ from __future__ import annotations
 import sys
 from copy import deepcopy
 from typing import Any
-from typing import Callable
-from typing import Dict
-from typing import List
-from typing import Optional
-from typing import Tuple
-from typing import Type
 from typing import TypeVar
 
-import z3
-
-from geneticengine.core.decorators import get_gengy
 from geneticengine.core.grammar import Grammar
 from geneticengine.core.random.sources import Source
 from geneticengine.core.representations.api import Representation
-from geneticengine.core.representations.tree.utils import GengyList
+from geneticengine.core.representations.tree.initializations import MethodType
+from geneticengine.core.representations.tree.initializations import mk_save_init
+from geneticengine.core.representations.tree.initializations import pi_grow_method
 from geneticengine.core.representations.tree.utils import relabel_nodes
 from geneticengine.core.representations.tree.utils import relabel_nodes_of_trees
 from geneticengine.core.tree import TreeNode
-from geneticengine.core.utils import build_finalizers
 from geneticengine.core.utils import get_arguments
 from geneticengine.core.utils import get_generic_parameter
 from geneticengine.core.utils import has_annotated_crossover
 from geneticengine.core.utils import has_annotated_mutation
 from geneticengine.core.utils import is_abstract
-from geneticengine.core.utils import is_annotated
-from geneticengine.core.utils import is_generic_list
-from geneticengine.core.utils import strip_annotations
 from geneticengine.exceptions import GeneticEngineError
-from geneticengine.metahandlers.base import is_metahandler
 
 T = TypeVar("T")
-
-
-def apply_metahandler(
-    r: Source,
-    g: Grammar,
-    receiver,
-    new_symbol,
-    depth: int,
-    ty: type[Any],
-    context: dict[str, str],
-) -> Any:
-    """
-    This method applies a metahandler to use a custom generator for things of a given type.
-    As an example, AnnotatedType[int, IntRange(3,10)] will use the IntRange.generate(r, recursive_generator)
-    The generator is the annotation on the type ("__metadata__").
-    """
-    metahandler = ty.__metadata__[0]
-    base_type = get_generic_parameter(ty)
-    return metahandler.generate(
-        r,
-        g,
-        receiver,
-        new_symbol,
-        depth,
-        base_type,
-        context,
-    )  # todo: last argument
-
-
-# TODO: make non static
-class SMTResolver:
-    clauses: list[Any] = []
-    receivers: dict[str, Callable] = {}
-    types: dict[str, Callable] = {}
-
-    @staticmethod
-    def add_clause(claus, recs: dict[str, Callable]):
-        SMTResolver.clauses.extend(claus)
-        for k, v in recs.items():
-            SMTResolver.receivers[k] = v
-
-    @staticmethod
-    def register_type(name, typ):
-        SMTResolver.types[name] = SMTResolver.to_z3_typ(typ)
-
-    @staticmethod
-    def to_z3_typ(typ):
-        return z3.Bool if typ == bool else z3.Int if typ == int else z3.Real
-
-    @staticmethod
-    def resolve_clauses():
-
-        if not SMTResolver.receivers:
-            return  # don't try to smt solve if we don't need to
-
-        solver = z3.Solver()
-
-        solver.set(":random-seed", 1)
-        solver.reset()
-
-        for clause in SMTResolver.clauses:
-            solver.add(clause(SMTResolver.types))
-        res = solver.check()
-
-        if res != z3.sat:
-            raise Exception(f"{solver} failed with {res}")
-
-        model = solver.model()
-        for (name, recv) in SMTResolver.receivers.items():
-            evaled = model.eval(
-                SMTResolver.types[name](
-                    name,
-                ),
-                model_completion=True,
-            )
-
-            recv(SMTResolver.get_type(evaled))
-
-        SMTResolver.clauses = []
-        SMTResolver.receivers = {}
-        SMTResolver.types = {}
-
-    @staticmethod
-    def get_type(evaled):
-        if type(evaled) == z3.z3.BoolRef:
-            evaled = bool(str(evaled))
-        elif type(evaled) == z3.z3.IntNumRef:
-            evaled = int(str(evaled))
-        elif type(evaled) == z3.z3.RatNumRef:
-            evaled = eval(str(evaled))
-        else:
-            raise NotImplementedError(
-                f"Don't know what to do with {type(evaled)} {evaled}",
-            )
-        return evaled
-
-    @staticmethod
-    def register_const(ident, val):
-        SMTResolver.register_type(ident, type(val))
-        ty = SMTResolver.types[ident]
-        SMTResolver.clauses.append(lambda _: ty(ident) == val)
-
-
-def Grow(
-    r: Source,
-    g: Grammar,
-    depth: int,
-    starting_symbol: type[Any] = int,
-):
-    """
-    Implements the standard Grow tree-initialization method, where trees are naturally grown from the grammar.
-    """
-
-    def filter_choices(possible_choices: list[type], depth):
-        valid_productions = [
-            vp for vp in possible_choices if g.get_distance_to_terminal(vp) <= depth
-        ]
-        return valid_productions
-
-    def handle_symbol(
-        next_type,
-        next_finalizer,
-        depth: int,
-        ident: str,
-        ctx: dict[str, str],
-    ):
-        expand_node(
-            r,
-            g,
-            handle_symbol,
-            filter_choices,
-            next_finalizer,
-            depth,
-            next_type,
-            ident,
-            ctx,
-        )
-
-    state = {}
-
-    def final_finalize(x):
-        state["final"] = x
-
-    handle_symbol(starting_symbol, final_finalize, depth, "root", ctx={})
-    SMTResolver.resolve_clauses()
-    n = state["final"]
-    relabel_nodes_of_trees(n, g)
-    return n
-
-
-def Full(
-    r: Source,
-    g: Grammar,
-    depth: int,
-    starting_symbol: type[Any] = int,
-):
-    """
-    Implements the Full tree-initialization method, where trees are grown from the grammar with all branches as deep as possible, making full trees.
-    """
-
-    def filter_choices(possible_choices: list[type], depth):
-        valid_productions = [
-            vp for vp in possible_choices if g.get_distance_to_terminal(vp) <= depth
-        ]
-        recursive_valid_productions = [
-            vp for vp in valid_productions if vp in g.recursive_prods
-        ]
-        if recursive_valid_productions:
-            return recursive_valid_productions
-        return valid_productions
-
-    def handle_symbol(
-        next_type,
-        next_finalizer,
-        depth: int,
-        ident: str,
-        ctx: dict[str, str],
-    ):
-        expand_node(
-            r,
-            g,
-            handle_symbol,
-            filter_choices,
-            next_finalizer,
-            depth,
-            next_type,
-            ident,
-            ctx,
-        )
-
-    state = {}
-
-    def final_finalize(x):
-        state["final"] = x
-
-    handle_symbol(starting_symbol, final_finalize, depth, "root", ctx={})
-    SMTResolver.resolve_clauses()
-    n = state["final"]
-    relabel_nodes_of_trees(n, g)
-    return n
-
-
-def Ramped_HalfAndHalf(
-    r: Source,
-    g: Grammar,
-    depth: int,
-    starting_symbol: type[Any] = int,
-):
-    """
-    Implements the Ramped Half and Half tree-initialization method, where trees are either grown through the Grow method or through the Full method.
-    """
-    methods = [Grow, Full]
-    method = r.choice(methods)
-    return method(r, g, depth, starting_symbol)
-
-
-def PI_Grow(
-    r: Source,
-    g: Grammar,
-    depth: int,
-    starting_symbol: type[Any] = int,
-):
-    """
-    Implements the PI Grow tree-initialization method (http://ncra.ucd.ie/papers/Exploring%20Position%20Independent%20Initialisation%20in%20Grammatical%20Evolution.pdf), where trees are grown to have at least one branch as deep as possible.
-    """
-    state = {}
-
-    def final_finalize(x):
-        state["final"] = x
-
-    prodqueue = []
-    nRecs = [0]
-
-    def handle_symbol(
-        next_type,
-        next_finalizer,
-        depth: int,
-        ident: str,
-        ctx: dict[str, str],
-    ):
-        prodqueue.append((next_type, next_finalizer, depth, ident, ctx))
-        if next_type in g.recursive_prods:
-            nRecs[0] += 1
-
-    handle_symbol(starting_symbol, final_finalize, depth, "root", ctx={})
-
-    def filter_choices(possible_choices: list[type], depth):
-        valid_productions = [
-            vp for vp in possible_choices if g.distanceToTerminal[vp] <= depth
-        ]
-        if (nRecs[0] == 0) and any(  # Are we the last recursive symbol?
-            [
-                prod in g.recursive_prods for prod in valid_productions
-            ],  # Are there any  recursive symbols in our expansion?
-        ):
-            valid_productions = [
-                vp for vp in valid_productions if vp in g.recursive_prods
-            ]  # If so, then only expand into recursive symbols
-
-        return valid_productions
-
-    while prodqueue:
-        next_type, next_finalizer, depth, ident, ctx = r.pop_random(prodqueue)
-        if next_type in g.recursive_prods:
-            nRecs[0] -= 1
-        expand_node(
-            r,
-            g,
-            handle_symbol,
-            filter_choices,
-            next_finalizer,
-            depth,
-            next_type,
-            ident,
-            ctx,
-        )
-    SMTResolver.resolve_clauses()
-    n = state["final"]
-    relabel_nodes_of_trees(n, g)
-    return n
 
 
 def random_node(
@@ -321,162 +29,18 @@ def random_node(
     g: Grammar,
     max_depth: int,
     starting_symbol: type[Any] = None,
-    method=Ramped_HalfAndHalf,
+    method: MethodType = pi_grow_method,
 ):
     if starting_symbol is None:
         starting_symbol = g.starting_symbol
     return method(r, g, max_depth, starting_symbol)
 
 
-def mk_save_init(starting_symbol: Any, receiver: Callable):
-    if isinstance(starting_symbol, type):
-        pass
-    elif isinstance(starting_symbol, GengyList):
-        starting_symbol = starting_symbol.new_like
-    else:
-        starting_symbol = type(starting_symbol)
-
-    def fin_recv(*x):
-        built = starting_symbol(*x)
-        built.gengy_init_values = x
-        return receiver(built)
-
-    return fin_recv
-
-
-def expand_node(
-    r: Source,
-    g: Grammar,
-    new_symbol,  # Method to find new_symbol (?)
-    filter_choices,
-    receiver,
-    depth,
-    starting_symbol: type,
-    id: str,
-    ctx: dict[str, str],
-) -> Any:
-    """
-    Creates a random node of a given type (starting_symbol)
-    """
-    if depth < 0:
-        raise GeneticEngineError("Recursion Depth reached")
-    if depth < g.get_distance_to_terminal(starting_symbol):
-        raise GeneticEngineError(
-            "There will be no depth sufficient for symbol {} in this grammar (provided: {}, required: {}).".format(
-                starting_symbol,
-                depth,
-                g.get_distance_to_terminal(
-                    starting_symbol,
-                ),
-            ),
-        )
-
-    if starting_symbol is int:
-        val = r.randint(0, 100, str(starting_symbol))
-        SMTResolver.register_const(id, val)
-        receiver(val)
-        return
-    elif starting_symbol is float:
-        valf = r.random_float(-100, 100, str(starting_symbol))
-        SMTResolver.register_const(id, valf)
-        receiver(valf)
-        return
-    elif starting_symbol is bool:
-        valb = r.random_bool(str(starting_symbol))
-        SMTResolver.register_const(id, valb)
-        receiver(valb)
-        return
-    elif is_generic_list(starting_symbol):
-        ctx = ctx.copy()
-        ctx["_"] = id
-        r.random_list(
-            receiver,
-            new_symbol,
-            depth,
-            starting_symbol,
-            ctx,
-            str(starting_symbol),
-        )
-        return
-    elif is_metahandler(starting_symbol):
-        ctx = ctx.copy()
-        ctx["_"] = id
-        apply_metahandler(
-            r,
-            g,
-            receiver,
-            new_symbol,
-            depth,
-            starting_symbol,
-            ctx,
-        )
-        return
-    else:
-        if starting_symbol not in g.all_nodes:
-            raise GeneticEngineError(
-                f"Symbol {starting_symbol} not in grammar rules.",
-            )
-
-        if starting_symbol in g.alternatives:  # Alternatives
-            extra_depth = 0
-            if is_abstract(starting_symbol) and g.expansion_depthing:
-                extra_depth = 1
-
-            compatible_productions = g.alternatives[starting_symbol]
-            valid_productions = filter_choices(
-                compatible_productions,
-                depth - extra_depth,
-            )
-            if not valid_productions:
-                raise GeneticEngineError(
-                    "No productions for non-terminal node with type: {} in depth {} (minimum required: {}).".format(
-                        starting_symbol,
-                        depth - extra_depth,
-                        str(
-                            [
-                                (vp, g.distanceToTerminal[vp])
-                                for vp in compatible_productions
-                            ],
-                        ),
-                    ),
-                )
-            if any(["weight" in get_gengy(p) for p in valid_productions]):
-                weights = [get_gengy(p).get("weight", 1.0) for p in valid_productions]
-                rule = r.choice_weighted(
-                    valid_productions,
-                    weights,
-                    str(starting_symbol),
-                )
-            else:
-                rule = r.choice(valid_productions, str(starting_symbol))
-            new_symbol(rule, receiver, depth - extra_depth, id, ctx)
-        else:  # Normal production
-            args = get_arguments(starting_symbol)
-            ctx = ctx.copy()
-            l: list[Any] = []
-            for argn, _ in args:
-                name = id + "_" + argn
-                ctx[argn] = name
-
-                def fn(val, name=name):
-                    pass
-
-                l.append(fn)
-
-            fins = build_finalizers(
-                mk_save_init(starting_symbol, receiver),
-                len(args),
-                l,
-            )
-            for i, (argn, argt) in enumerate(args):
-                new_symbol(argt, fins[i], depth - 1, id + "_" + argn, ctx)
-
-
 def random_individual(
     r: Source,
     g: Grammar,
     max_depth: int = 5,
-    method=Ramped_HalfAndHalf,
+    method: MethodType = pi_grow_method,
 ) -> TreeNode:
     try:
         assert max_depth >= g.get_min_tree_depth()
@@ -488,7 +52,7 @@ def random_individual(
         raise GeneticEngineError(
             f"Cannot use complete grammar for individual creation. Max depth ({max_depth}) is smaller than grammar's minimal tree depth ({g.get_min_tree_depth()}).",
         )
-    ind = random_node(r, g, max_depth, g.starting_symbol, method=method)
+    ind = random_node(r, g, max_depth, g.starting_symbol, method)
     assert isinstance(ind, TreeNode)
     return ind
 
@@ -532,14 +96,13 @@ def mutate_inner(
                     random_node,
                     max_depth - 1,
                     get_generic_parameter(arg_to_be_mutated),
-                    method=Grow,
                     current_node=args[index],
                 )
                 mk = mk_save_init(type(i), lambda x: x)(*args)
                 return mk
 
             for _ in range(5):
-                replacement = random_node(r, g, max_depth, ty, method=Grow)
+                replacement = random_node(r, g, max_depth, ty)
                 if replacement != i:
                     break
             return replacement
@@ -575,7 +138,7 @@ def mutate_inner(
             return mk
     else:
         for _ in range(5):
-            rn = random_node(r, g, max_depth, ty, method=Grow)
+            rn = random_node(r, g, max_depth, ty)
             if rn != i:
                 break
         return rn
@@ -777,7 +340,7 @@ def crossover_inner(
                 replacement = r.choice(options)
             if replacement is None:
                 for _ in range(5):
-                    replacement = random_node(r, g, max_depth, ty, method=Grow)
+                    replacement = random_node(r, g, max_depth, ty)
                     if replacement != i:
                         break
 
@@ -970,13 +533,8 @@ class TreeBasedRepresentation(Representation[TreeNode]):
     """This class represents the tree representation of an individual.
     In this approach, the genotype and the phenotype are exactly the same."""
 
-    method: Callable[[Source, Grammar, int, type[Any]], Any]
-
-    def __init__(self, method=Ramped_HalfAndHalf) -> None:
-        self.method = method
-
     def create_individual(self, r: Source, g: Grammar, depth: int) -> TreeNode:
-        return random_individual(r, g, depth, self.method)
+        return random_individual(r, g, depth)
 
     def mutate_individual(
         self,
