+++ tests/depthing_test.py
@@ -8,7 +8,7 @@ from typing import List
 from geneticengine.core.grammar import extract_grammar
 from geneticengine.core.grammar import Grammar
 from geneticengine.core.random.sources import RandomSource
-from geneticengine.core.representations.tree.treebased import PI_Grow
+from geneticengine.core.representations.tree.treebased import pi_grow_method
 from geneticengine.core.representations.tree.treebased import random_node
 from geneticengine.core.representations.tree.utils import GengyList
 from geneticengine.metahandlers.ints import IntRange
@@ -48,13 +48,13 @@ class TestDepthing:
     def test_normal_depthing(self):
         r = RandomSource(seed=1)
         g: Grammar = extract_grammar([Concrete], Root, False)
-        x = random_node(r, g, 4, Root, method=PI_Grow)
+        x = random_node(r, g, 4, Root, method=pi_grow_method)
         g: Grammar = extract_grammar([Concrete, Middle], Root, False)
-        y = random_node(r, g, 4, Root, method=PI_Grow)
+        y = random_node(r, g, 4, Root, method=pi_grow_method)
         g: Grammar = extract_grammar([Concrete, Middle, MiddleList], Root, False)
-        z = random_node(r, g, 4, Root, method=PI_Grow)
+        z = random_node(r, g, 4, Root, method=pi_grow_method)
         g: Grammar = extract_grammar([ConcreteList, Middle], Root, False)
-        a = random_node(r, g, 4, Root, method=PI_Grow)
+        a = random_node(r, g, 4, Root, method=pi_grow_method)
         assert x.gengy_distance_to_term == 1
         assert x.gengy_nodes == 1
         assert x.gengy_weighted_nodes == 1
@@ -71,11 +71,11 @@ class TestDepthing:
     def test_expansion_depthing(self):
         r = RandomSource(seed=1)
         g: Grammar = extract_grammar([Concrete], Root, True)
-        x = random_node(r, g, 4, Root, method=PI_Grow)
+        x = random_node(r, g, 4, Root, method=pi_grow_method)
         g: Grammar = extract_grammar([Concrete, Middle], Root, True)
-        y = random_node(r, g, 8, Root, method=PI_Grow)
+        y = random_node(r, g, 8, Root, method=pi_grow_method)
         g: Grammar = extract_grammar([Concrete, Middle, MiddleList], Root, True)
-        z = random_node(r, g, 8, Root, method=PI_Grow)
+        z = random_node(r, g, 8, Root, method=pi_grow_method)
         assert x.gengy_distance_to_term == 2
         assert x.gengy_nodes == 2
         assert x.gengy_weighted_nodes == 3
