+++ tests/grammar_test.py
@@ -7,8 +7,8 @@ from typing import Annotated
 from typing import List
 from typing import Type
 from unittest import skip
-import pytest
 
+import pytest
 from scipy import rand
 
 from geneticengine.algorithms.gp.gp import GP
@@ -17,8 +17,6 @@ from geneticengine.core.decorators import abstract
 from geneticengine.core.grammar import extract_grammar
 from geneticengine.core.grammar import Grammar
 from geneticengine.core.random.sources import RandomSource
-from geneticengine.core.representations.tree.treebased import Grow
-from geneticengine.core.representations.tree.treebased import PI_Grow
 from geneticengine.core.representations.tree.treebased import random_node
 from geneticengine.core.utils import get_arguments
 from geneticengine.metahandlers.ints import IntRange
@@ -75,7 +73,7 @@ class TestGrammar:
     def test_rec(self):
         r = RandomSource(seed=1)
         g: Grammar = extract_grammar([Leaf, Rec], Root)
-        x = random_node(r, g, 10, Root, method=PI_Grow)
+        x = random_node(r, g, 10, Root)
         # print(x) -- Leaf()
         assert isinstance(x, Rec)
         assert isinstance(x, Root)
@@ -88,7 +86,6 @@ class TestGrammar:
             g,
             max_depth=15,
             starting_symbol=Root,
-            method=PI_Grow,
         )
         assert contains_type(x, RecAlt)
         assert isinstance(x, Root)
@@ -102,7 +99,6 @@ class TestGrammar:
             g,
             max_depth=2,
             starting_symbol=Root,
-            method=PI_Grow,
         )
         assert isinstance(x, Leaf)
         assert isinstance(x, Root)
