+++ tests/ind_gen_test.py
@@ -8,9 +8,9 @@ from typing import List
 from geneticengine.core.grammar import extract_grammar
 from geneticengine.core.grammar import Grammar
 from geneticengine.core.random.sources import RandomSource
-from geneticengine.core.representations.tree.treebased import Full
-from geneticengine.core.representations.tree.treebased import PI_Grow
-from geneticengine.core.representations.tree.treebased import Ramped_HalfAndHalf
+from geneticengine.core.representations.tree.initializations import full_method
+from geneticengine.core.representations.tree.initializations import grow_method
+from geneticengine.core.representations.tree.initializations import pi_grow_method
 from geneticengine.core.representations.tree.treebased import random_node
 from geneticengine.core.representations.tree.utils import GengyList
 from geneticengine.metahandlers.ints import IntRange
@@ -62,28 +62,28 @@ class TestPIGrow:
     def test_root(self):
         r = RandomSource(seed=1)
         g: Grammar = extract_grammar([Concrete], Root)
-        x = random_node(r, g, 4, Root, method=PI_Grow)
+        x = random_node(r, g, 4, Root, method=pi_grow_method)
         assert isinstance(x, Concrete)
         assert isinstance(x, Root)
 
     def test_leaf(self):
         r = RandomSource(seed=1)
         g: Grammar = extract_grammar([Leaf], Root)
-        x = random_node(r, g, 4, Leaf, method=PI_Grow)
+        x = random_node(r, g, 4, Leaf, method=pi_grow_method)
         assert isinstance(x, Leaf)
         assert isinstance(x, Root)
 
     def test_leaf2(self):
         r = RandomSource(seed=1)
         g: Grammar = extract_grammar([Concrete], Root)
-        x = random_node(r, g, 4, Concrete, method=PI_Grow)
+        x = random_node(r, g, 4, Concrete, method=pi_grow_method)
         assert isinstance(x, Concrete)
         assert isinstance(x, Root)
 
     def test_concrete_list(self):
         r = RandomSource(seed=1)
         g: Grammar = extract_grammar([ConcreteList], Root)
-        x = random_node(r, g, 6, Root, method=PI_Grow)
+        x = random_node(r, g, 6, Root, method=pi_grow_method)
         assert isinstance(x, ConcreteList)
         assert isinstance(x.xs, list)
         assert isinstance(x, Root)
@@ -91,7 +91,7 @@ class TestPIGrow:
     def test_concrete_annotated_list(self):
         r = RandomSource(seed=1)
         g: Grammar = extract_grammar([ConcreteAnnotatedList], Root)
-        x = random_node(r, g, 6, Root, method=PI_Grow)
+        x = random_node(r, g, 6, Root, method=pi_grow_method)
         assert isinstance(x, ConcreteAnnotatedList)
         assert isinstance(x.xs, list)
         assert isinstance(x, Root)
@@ -99,7 +99,7 @@ class TestPIGrow:
     def test_middle_list(self):
         r = RandomSource(seed=1)
         g: Grammar = extract_grammar([MiddleList, Concrete], Root)
-        x = random_node(r, g, 6, Root, method=PI_Grow)
+        x = random_node(r, g, 6, Root, method=pi_grow_method)
         assert isinstance(x, MiddleList)
         assert isinstance(x.z, list)
         assert isinstance(x, Root)
@@ -111,53 +111,27 @@ class TestPIGrow:
 
         r = RandomSource(seed=1)
         g: Grammar = extract_grammar([Concrete, Middle, Root], RootHolder)
-        x = random_node(r, g, 20, RootHolder, method=PI_Grow)
+        x = random_node(r, g, 20, RootHolder, method=pi_grow_method)
         assert x.gengy_distance_to_term == 20
         assert isinstance(x, RootHolder)
 
 
-class TestFull:
+class TestFullMethod:
     def test_root(self):
         r = RandomSource(seed=1)
         g: Grammar = extract_grammar([Concrete], Root)
-        x = random_node(r, g, 4, Root, method=Full)
+        x = random_node(r, g, 4, Root, method=full_method)
         assert isinstance(x, Concrete)
         assert isinstance(x, Root)
 
     def test_middle_double(self):
         r = RandomSource(seed=1)
         g: Grammar = extract_grammar([Concrete, MiddleDouble], Root)
-        x = random_node(r, g, 4, Root, method=Full)
+        x = random_node(r, g, 4, Root, method=full_method)
         assert x.gengy_nodes == 15
-        y = random_node(r, g, 5, Root, method=Full)
+        y = random_node(r, g, 5, Root, method=full_method)
         assert y.gengy_nodes == 31
-        z = random_node(r, g, 6, Root, method=Full)
+        z = random_node(r, g, 6, Root, method=full_method)
         assert z.gengy_nodes == 63
-        x1 = random_node(r, g, 7, Root, method=Full)
+        x1 = random_node(r, g, 7, Root, method=full_method)
         assert x1.gengy_nodes == 127
-
-
-class TestRamped:
-    def test_root(self):
-        r = RandomSource(seed=1)
-        g: Grammar = extract_grammar([Concrete], Root)
-        x = random_node(r, g, 4, Root, method=Ramped_HalfAndHalf)
-        assert isinstance(x, Concrete)
-        assert isinstance(x, Root)
-
-    def test_middle_double(self):
-        r = RandomSource(seed=1)
-        g: Grammar = extract_grammar([Concrete, MiddleDouble], Root)
-        individuals = []
-        depths = []
-        nodes = []
-        for _ in range(100):
-            x = random_node(r, g, 9, Root, method=Ramped_HalfAndHalf)
-            individuals.append(x)
-            depths.append(x.gengy_distance_to_term)
-            nodes.append(x.gengy_nodes)
-
-        assert max(depths) == 9
-        assert depths.count(max(depths)) > 45 and depths.count(max(depths)) < 60
-        assert max(nodes) == 511
-        assert nodes.count(max(nodes)) > 45 and nodes.count(max(nodes)) < 55
