+++ tests/is_terminal_test.py
@@ -7,7 +7,7 @@ from typing import Annotated
 from geneticengine.core.grammar import extract_grammar
 from geneticengine.core.grammar import Grammar
 from geneticengine.core.random.sources import RandomSource
-from geneticengine.core.representations.tree.treebased import PI_Grow
+from geneticengine.core.representations.tree.treebased import pi_grow_method
 from geneticengine.core.representations.tree.treebased import random_node
 from geneticengine.core.representations.tree.utils import GengyList
 from geneticengine.core.representations.tree.utils import relabel_nodes
